Add support for reading- and writing less than the object size.
This allows you to read- and write structs as uint8_t buffers without having to
go through an intermediate buffer or a union.

New functions have been introduced to accept a size variable:
- ringfs_append_ex
- ringfs_fetch_ex

The original functions were changed to invoke the _ex versions with the object
size as the read/write size. This was done to maintain source compatibility.